Finish multiple choice quiz and refactor multiple choice page
