+++ src/components/resultCard.js
@@ -1,35 +1,23 @@
 import React from 'react';
-import { Link } from 'react-router-dom';
 
 import {
   QuestionCont,
   Img,
-  ImgCont,
-  QuestionBtn,
+  ImgCont
 } from './../styled-components/Cards';
 
-import { useAuth } from './../contexts/AuthContext';
-import { updateScore } from './../database/queries';
 
-export default function ResultCard({ imgSrc, text }) {
-  const { currentUser } = useAuth();
+
+export default function ResultCard({ imgSrc, text, children }) {
+ 
 
   return (
     <QuestionCont>
       <Img src={imgSrc} />
       <ImgCont background={'#08302e'}>
         {text}
+        {children}
         <br />
-        <Link to="/account">
-          <QuestionBtn
-            onClick={() => {
-              updateScore(currentUser.uid);
-            }}
-            background={'#08302e'}
-          >
-            Account
-          </QuestionBtn>
-        </Link>
       </ImgCont>
     </QuestionCont>
   );
