+++ src/pages/MultipleChoicePage.js
@@ -1,14 +1,18 @@
 import React, { useEffect, useState } from 'react';
 import {
-  QuizContainer,
-  ImgCont,
-  QuestionCont,
-  QuestionBtn,
-  Img,
+  // QuizContainer,
+  // ImgCont,
+  // QuestionCont,
+  QuestionBtn
+  // Img,
 } from '../styled-components/Cards';
 
 import { db } from './../database/firebase';
+import { Link, useLocation } from 'react-router-dom';
+import { useAuth } from './../contexts/AuthContext';
+import { updateScore } from './../database/queries';
 
+import MultipleChoiceCard from './../components/multipleChoice';
 import DescriptionCard from './../components/descriptionCard';
 import ResultCard from './../components/resultCard';
 
@@ -16,27 +20,47 @@ import loser from './../images/loser.png';
 import winner from './../images/winner.png';
 
 export default function MultipleChoice() {
+
+  const { currentUser } = useAuth();
+
   const [data, setData] = useState([]);
   const [question, setQuestion] = useState(0);
   const [colour, setColour] = useState('transparent');
   const [toggle, setToggle] = useState(false);
-  const [what, setWhat] = useState();
+  const [answer, setAnswer] = useState();
   const [win, setWin] = useState(true);
-  // const [score, setScore] = useState(0);
+  const [stage, setStage] = useState('key-stage-3');
+  const [timer,setTimer] = useState(false);
 
-  const docRef = db
+  
+const location = useLocation();
+
+if(stage && location.pathname.includes('k4') ){
+  setStage('key-stage-4')
+}
+
+let docRef;
+
+if(stage){
+  docRef = db
     .collection('Quizzes')
-    .doc('key-stage-3')
+    .doc(stage)
     .collection('multiple-choice');
+}
+   
 
   const collectionArr = [];
 
   useEffect(() => {
-    docRef
+    const timer = setTimeout(() => {
+      setTimer(true);
+    }, 3000);
+
+    if(stage) {
+      docRef
       .get()
       .then((result) => {
         return result.docs.forEach((doc) => {
-          // console.log('doc', doc.data());
           collectionArr.push(doc.data());
         });
       })
@@ -44,59 +68,30 @@ export default function MultipleChoice() {
         console.log('array', collectionArr);
         return setData(collectionArr);
       })
-      .catch((error) => console.log(error));
-  }, []);
+      .catch((error) => console.log(error));}
+      return () => clearTimeout(timer);
+    }, []);
+    
 
-  function correctAnswer(event) {
-    const check = data[question].options.filter(
-      (e) => e.text === event.target.textContent
-    );
-    console.log('check', check[0].answer);
-
-    if (check[0].answer === false) {
-      setToggle(true);
-      setColour('red');
-      setWhat(check[0]);
-      setWin(false);
-    } else {
-      setToggle(true);
-      setColour('green');
-      setWhat(check[0]);
-    }
-  }
 
   return (
     <div className="mc">
       {data[question] ? (
         <>
           {!toggle ? (
-            <QuizContainer>
-              <ImgCont>
-                <ImgCont width="100%" height="10%">
-                  {data[question].question}
-                </ImgCont>
-                <Img src={data[question].image} alt="question img" />
-              </ImgCont>
-              <QuestionCont>
-                {data[question].options.map((e) => (
-                  <>
-                    <QuestionBtn
-                      onClick={correctAnswer}
-                      background={colour}
-                      key={data[question].options.text}
-                    >
-                      {e.text}
-                    </QuestionBtn>
-                    <br />
-                  </>
-                ))}
-              </QuestionCont>
-              {/* <QuestionBtn></QuestionBtn> */}
-            </QuizContainer>
+           <MultipleChoiceCard
+           colour={colour}
+           data={data}
+           question={question}
+           setToggle={setToggle}
+           setColour={setColour}
+           setAnswer={setAnswer}
+           setWin={setWin}
+           />
           ) : (
             <DescriptionCard
               background={colour}
-              answerObj={what}
+              answerObj={answer}
               question={question}
               setQuestion={setQuestion}
               toggle={toggle}
@@ -106,14 +101,27 @@ export default function MultipleChoice() {
           )}
         </>
       ) : question > collectionArr.length - 1 ? (
-        win ? (
-          <ResultCard imgSrc={winner} text={'better luck next time!! '} />
-        ) : (
-          <ResultCard imgSrc={loser} text={'better luck next time!! '} />
-        )
+        win && timer ?  (
+          <ResultCard imgSrc={winner} text={'Congrats!! '}>
+
+          <Link to="/account">
+          <QuestionBtn
+            onClick={() => {
+              updateScore(currentUser.uid);
+            }}
+            background={'#08302e'}
+          >
+            Click Here to Save progress !
+          </QuestionBtn>
+        </Link>
+          </ResultCard>
+        ) : !win && timer ?(          
+          <ResultCard imgSrc={loser} text={'Better luck next time!! '} ><Link to="/account">Account</Link></ResultCard> 
+        ):null
       ) : (
         'null'
       )}
     </div>
   );
 }
+
